+++ raytracer/features/Tuple/Point.cpp
@@ -4,7 +4,7 @@
 
 #include "Point.h"
 
-Point::Point(float x, float y, float z) {
+Point::Point(double x, double y, double z) {
     this->x = x;
     this->y = y;
     this->z = z;
