+++ raytracer/features/Tuple/Vector.cpp
@@ -4,7 +4,7 @@
 
 #include "Vector.h"
 
-Vector::Vector(float x, float y, float z) {
+Vector::Vector(double x, double y, double z) {
     this->x = x;
     this->y = y;
     this->z = z;
